remove tag_size_empty size only if new item is actually created

+++ app/admin/item.rb
@@ -24,18 +24,18 @@ ActiveAdmin.register Item do
     end
 
     def create
-      # raise params[:item][:garment_id]
-
-      if Garment.find(params[:item][:garment_id]).tag_size_empty.present?
-        empty_sizes = Garment.find(params[:item][:garment_id]).tag_size_empty.gsub(' ', '').split(',')
-        new_size = params[:item][:tag_size].gsub(' ', '').split(',')
-        new_empty_sizes = (empty_sizes - new_size).join(', ')
-
-        Garment.find(params[:item][:garment_id]).update_attributes(tag_size_empty: new_empty_sizes)
-      end
-
       create! do |format|
-        format.html { redirect_to session.delete(:return_to) }
+        if @item.save
+          if Garment.find(params[:item][:garment_id]).tag_size_empty.present?
+            empty_sizes = Garment.find(params[:item][:garment_id]).tag_size_empty.gsub(' ', '').split(',')
+            new_size = params[:item][:tag_size].gsub(' ', '').split(',')
+            new_empty_sizes = (empty_sizes - new_size).join(', ')
+
+            Garment.find(params[:item][:garment_id]).update_attributes(tag_size_empty: new_empty_sizes)
+          end
+
+          format.html { redirect_to session.delete(:return_to) }
+        end
       end
     end
